SQl Connection Pool

+++ httpServer/src/commonMain/kotlin/pw/binom/io/httpServer/HttpResponse.kt
@@ -12,7 +12,7 @@ interface HttpResponse {
     fun resetHeader(name: String, value: String)
     fun addHeader(name: String, value: String)
     var enableKeepAlive:Boolean
-    suspend fun complete(autoFlushSize:UInt= DEFAULT_BUFFER_SIZE.toUInt()):HttpResponseBody
+    suspend fun complete(autoFlushSize: Int = DEFAULT_BUFFER_SIZE): HttpResponseBody
 }
 
 interface HttpResponseBody:AsyncOutput{
